requirements.txt
alabaster==0.7.12
albumentations==0.5.2
awscli==1.18.209
Babel==2.9.0
backcall==0.2.0
botocore==1.19.49
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.3
coverage==5.3.1
cycler==0.10.0
decorator==4.4.2
docutils==0.16
flake8==3.8.4
gdcm==1.1
idna==2.10
imageio==2.9.0
imagesize==1.2.0
imgaug==0.4.0
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.0.0
jupyter-client==6.1.7
jupyter-core==4.7.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
multipart==0.2.3
networkx==2.5
numpy==1.19.5
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
packaging==20.8
pandas==1.2.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pluggy==0.13.1
progressbar2==3.43.1
prompt-toolkit==3.0.9
ptyprocess==0.7.0
pyasn1==0.4.8
pycodestyle==2.6.0
pydicom==2.1.2
pyflakes==2.2.0
Pygments==2.7.3
pylibjpeg==1.1.1
pylibjpeg-libjpeg==1.1.0
pylibjpeg-openjpeg==1.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
python-utils==2.4.0
pytz==2020.5
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==20.0.0
requests==2.25.1
rsa==4.5
s3transfer==0.3.3
scikit-image==0.18.1
scikit-learn==0.24.0
scipy==1.6.0
Shapely==1.7.1
six==1.15.0
sklearn==0.0
snowballstemmer==2.0.0
Sphinx==3.4.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
threadpoolctl==2.1.0
tifffile==2020.12.8
torch==1.7.1
torchvision==0.8.2
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
ujson==4.0.1
urllib3==1.26.2
wcwidth==0.2.5
